Create leeches_MDedit4.ESP
remove blood from the first victim in sewer

leaner doesn't have a name and the lantern that was in front of his face is gone for some reason

door to sewer doesn't require the key. can be unlocked and entered normally. needs script saying it requires a key or something

detective's office needs to be marked as illegal to rest

rename slough crawlers to rotcreepers

first rotcreeper encounter needs collision blocks to make him move around reliably

cultist lair is red and foggy

connect cave to sewer